requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
python-dotenv==0.19.0
Werkzeug==2.0.1
requests==2.26.0
PyJWT==2.1.0
python-dateutil==2.8.2
reportlab==4.4.0
qrcode==7.3
Pillow==10.2.0
Flask-Migrate==3.1.0
